Fixed some warnings from PGI compiler dashboard
Mostly dead code, unused variables, or type
conversion warnings.

+++ Common/Misc/vtkHeap.cxx
@@ -19,7 +19,7 @@
 
 vtkStandardNewMacro(vtkHeap);
 
-static int vtkGetLongAlignment()
+static size_t vtkGetLongAlignment()
 {
   struct vtkTestAlignLong
   {
